requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
alipay-sdk-python==3.7.868
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
alphashape==1.3.1
altair==6.0.0
anthropic==0.84.0
anyio==4.12.1
archspec==0.1.2
asgiref==3.8.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
boltons==23.0.0
boto3==1.42.59
botocore==1.42.59
brotli==1.2.0
cachetools==6.2.6
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.3.0
click-log==0.4.0
colorama==0.4.6
conda==23.11.0
conda-anaconda-telemetry==0.1.3
conda-anaconda-tos==0.7.1
conda-content-trust==0.2.0
conda-libmamba-solver==24.7.1
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
Crawl4AI==0.8.0
crcmod==1.7
cryptography==42.0.8
cssselect==1.3.0
dashscope==1.25.10
dataclasses-json==0.6.7
django==5.2.8
django-cors-headers==4.7.0
django-filter==25.1
django-ipware==7.0.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
docstring_parser==0.17.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.134.0
fastuuid==0.14.0
filelock==3.20.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozendict==2.4.0
frozenlist==1.8.0
fsspec==2025.12.0
GitPython==3.1.46
greenlet==3.3.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.3
humanize==4.15.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==0.3.25
langchain-classic==1.0.1
langchain-community==0.3.25
langchain-core==0.3.83
langchain-experimental==0.3.4
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.3.45
lark==1.3.1
litellm==1.81.3
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
mcp==1.25.0
mdurl==0.1.2
menuinst==2.1.2
mpmath==1.3.0
multidict==6.7.1
mypy-extensions==1.1.0
mysql-connector-python==9.6.0
mysqlclient==2.2.7
narwhals==2.17.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.5
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.2
oss2==2.19.1
packaging==25.0
pandas==2.3.3
patchright==1.57.2
pillow==11.2.1
platformdirs==4.3.2
playwright==1.52.0
pluggy==1.5.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.1
pyarrow==23.0.1
pyasn1==0.6.2
pycosat==0.6.6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.27.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==25.3.0
pypinyin==0.54.0
PySocks==1.8.1
pytest==9.0.2
python-a2a==0.5.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-ipware==3.0.0
python-multipart==0.0.21
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
redis==7.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.16.0
safetensors==0.7.0
schedule==1.2.2
scipy==1.17.0
setuptools==78.1.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
skills-for-all-agent==0.1.1
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlparse==0.5.3
sse-starlette==3.2.0
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tornado==6.5.4
tqdm==4.66.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
uuid_utils==0.14.0
uvicorn==0.40.0
watchdog==6.0.0
websocket-client==1.9.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.23.0